python_requirement.txt
absl-py==0.7.1
asn1crypto==0.24.0
astor==0.7.1
certifi==2019.6.16
cffi==1.12.2
chardet==3.0.4
conda==4.7.10
conda-package-handling==1.3.11
cryptography==2.6.1
cycler==0.10.0
decorator==4.4.0
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
idna==2.8
imageio==2.5.0
imgaug==0.4.0
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
libarchive-c==2.8
lime==0.1.1.36
Markdown==3.1
matplotlib==3.1.1
menuinst==1.4.16
mkl-fft==1.0.12
mkl-random==1.0.2
munch==2.5.0
networkx==2.3
numpy==1.16.2
olefile==0.46
opencv-python==4.1.2.30
Pillow==6.1.0
pretrainedmodels==0.7.4
protobuf==3.7.1
pycosat==0.6.3
pycparser==2.19
pyOpenSSL==19.0.0
pyparsing==2.4.0
PySocks==1.6.8
pyspark==2.3.2
python-dateutil==2.8.0
PyWavelets==1.0.3
pywin32==223
PyYAML==5.1
requests==2.21.0
ruamel-yaml==0.15.46
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.2.1
Shapely==1.7.0
six==1.12.0
tensorboard==1.12.2
tensorflow-gpu==1.12.0
termcolor==1.1.0
torch==1.1.0
torchvision==0.3.0
tqdm==4.32.1
urllib3==1.24.1
Werkzeug==0.15.2
win-inet-pton==1.1.0
wincertstore==0.2
